キャストに失敗する部分を修正

+++ FireFly/ViewController.swift
@@ -41,7 +41,7 @@ class ViewController: UIViewController ,CLLocationManagerDelegate{
         for  beacon in beacons as CLBeacon[]{
             println(beacon)
             
-            UIScreen.mainScreen().brightness = Float(beacon.accuracy);
+            UIScreen.mainScreen().brightness = CGFloat(beacon.accuracy);
         }
     }
